Added safety measures.
The macro for generating the options did not account for rng's min <=
max rule. This commit makes sure to make settings abide by this rule
within the macro.

For floats, the value was incremented by 10e-10_f64. This is because
f64::MIN_POSITIVE didn't work.

In addition, certain mutations such as Loop, Swap, and Multiply, were
made to work correctly and avoid crashing the program.